:sparkles: :white_check_mark: enable changes for object graphs that are DAGS, and not just trees.
The general structure of this change is to 1) store pending writes to
objects in a WeakMap so that changes are keyed by object reference
instead of qualified property name, and 2) do a DFS walk of the entire
object graph when merging the changes instead of only recursing on
changed paths. The central challenge of operating on a DAG is that
there's no easy way to tell what an object's parents are when merging
changes into the object; I tackeled this by using two lookup tables,
(one to cache the results of applying the changes) so that for any
object reference in the source graph we could either apply the changes
ourselves or pull the cached result.